Added default implementation to CausableGraph and CausableGraphReasoning traits.
